convert statistics to long

+++ main.h
@@ -130,15 +130,15 @@ struct history {
 extern struct history hist;
 
 struct statistics {
-	unsigned int packets;
-	unsigned int bytes;
-	unsigned int airtimes;
-	unsigned int packets_per_rate[MAX_RATES];
-	unsigned int bytes_per_rate[MAX_RATES];
+	unsigned long packets;
+	unsigned long bytes;
+	unsigned long airtimes;
+	unsigned long packets_per_rate[MAX_RATES];
+	unsigned long bytes_per_rate[MAX_RATES];
 
-	unsigned int packets_per_type[MAX_FSTYPE];
-	unsigned int bytes_per_type[MAX_FSTYPE];
-	unsigned int airtime_per_type[MAX_FSTYPE];
+	unsigned long packets_per_type[MAX_FSTYPE];
+	unsigned long bytes_per_type[MAX_FSTYPE];
+	unsigned long airtime_per_type[MAX_FSTYPE];
 };
 
 extern struct statistics stat;
